moved pages from components to components/pages
