+++ react-frontend/src/components/Pages/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './Header.js';
-import Footer from './Footer.js';
-import HorizontalScrollList from './HorizontalScrollList.js';
+import Header from '../Header.js';
+import Footer from '../Footer.js';
+import HorizontalScrollList from '../HorizontalScrollList.js';
 import { useState, useEffect } from 'react';
