+++ react-frontend/src/components/Pages/SignIn.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './Header.js';
-import Footer from './Footer.js';
+import Header from '../Header.js';
+import Footer from '../Footer.js';
 
 const SignIn = () => {
   return (
